QScriptEngine::setGlobalObject set the prototype
setGlobalObject will actually set the prototype of the global object

The reason we do that is that if we set an interceptor, all the properteis
will be intercepted.

But we have to work around the fact that for v8, the real global
object is the prototype of the global object.  So we use the prototype
of the global object everywhere.