Improve Renderer::preprocess, NodeUpdater and general pluggability
Fixed the preprocess so that it doesn't do two explicit update passes, one before
preprocess and one after, to be in sync with *blocked* nodes. Instead introduce
a virtual isNodeBlocked() in NodeUpdater so that node updaters can use the same
mechanics for a single node as it otherwise would have with the full pass.

Also make the NodeUpdater pluggable in renderer, in case someone would like
to install their own updater which does something special..