PathView items were not correctly updated when rootIndex changed.
PathView got confused because it queried the model when removing
items due to the rootIndex change. This caused new items to be added
since the model->count() was not zero (it was the count for the new
rootIndex).  Calculate the PathView modelCount similarly to the other
views (rather than querying the model) and correctly handle the
modelCount being set to zero.

Change-Id: I9e4fb27c14aa8c77ad37ff01d95c123f909db20d
Task-number: QTBUG-16366
Reviewed-by: Bea Lam