QtDBus meta object: keep methods, signals and props with unknown types
Use VoidStar as the metatype, with an unknown type called
"QDBusRawType". The actual D-Bus type is saved as an hex value as a
template parameter.

  D-Bus type	Qt type			comment
    h		QDBusRawType<0x68>*	Unix file descriptors
    ~		QDBusRawType<0x7e>*	invalid type
    ai		QDBusRawType<0x6169>*	array of int32
    a{i(ssy)}	QDBusRawType<0x617b6928737379297d>*

Note that the number in the template doesn't have to be valid. The
QDBusRawType class doesn't exist anyway.

I thought of just leaving the raw D-Bus type there, but who knows what
kind of things can appear there, like other '>' (which may cause
problems for anything trying to parse the meta object later).

Task-number: QTBUG-17476