Make addImportPath() work for windows paths starting with lower case
Was causing assert failure on windows if the 'c:' was lower case,
since it was being added to the import path database with a lower
case and thus later lookups with an upper case 'c:' would fail.
This change fixes the check for whether the path refers to a
local path or not.

Task-number: QTBUG-16885
Change-Id: I0a2a2f705443ed453fb2b13f8599e035c2bd2877
Reviewed-by: Martin Jones