Flickable uses the flick velocity to determine whether to retain grab
Flickable keeps the mouse grab if it was recently flicked and another
flick is started before it has settled. However, it is using the
velocity of the flick rather than the instantaneous velocity of the
view, which causes it to be grabbed unless the view has come to a
complete stop.  Use smoothedVelocity which is updated during the view
movement. Also increase the threshold a little.

Change-Id: I970318680d38103468155fa566c489c7874d1b00
Task-number: QTBUG-17383
Reviewed-by: Michael Brasser