Correct the "module not installed" error handling
The not installed error will be issued if, after loading plugins and considering
the contents of qmldir, there are no elements in the particular minor version.

For example, if a plugin/qmldir provides the following types (either from C++ or
as QML files specified in the qmldir file):

Foo 1.1
Bar 1.3

importing versions 1.0, 1.2, or 1.4 will fail with the not installed error.

Change-Id: I8566fda6918cb48936144e67a1ce75add0f160d8
Task-number: QTBUG-17324
Reviewed-by: Martin Jones