Cocoa/Alien: bugfix corner cases with popups and enter/leave
It turns out that we sometimes hit a strange bug with enter/leave
events when a popup is showing. If you righpress to show the popup,
and then move the mouse outside the window, we get a continues series
of leave events. This patch separates more the native vs alien logic
for dispatching enter/leave to accommondate this problem