Add a set of connection capabilities to QDBusConnection
The capabilities are negotiated with the D-Bus peer or bus at
connection time and may include extra features not available when
D-Bus 1.0 was released.

Currently (as of D-Bus 1.4), the only additional feature is Unix file
descriptor passing.

Proposed features are "maybe types" and single-precision floating
point. They may be added to D-Bus 1.6.

Task-number: QTBUG-17478