Keep sending posted events after commit 7196045b78b33cf135683d5c0b4e164f95231791
This commit put in place a timer based mechanism to ensure that posted
events continue to be sent. As pointed out in commit
eb1015c7bbf135af3656110a4d112377c1209db8, GetQueueStatus() can indicate
that there are messages in the queue but our hook never actually
receives them, hence the timer solution.

However, having the message type and hwnd "guards" in the conditional
for starting a timer would cause us to lose events in some situations,
due to the fact that QEventDispatcherWin32::processEvents() will only
allow one WM_QT_SENDPOSTEDEVENTS message per call (multiple messages
will be compressed into a single one). When racing with other threads,
it would be possible to end up blocking one thread that has pending
posted events unintentionally.

This fix removes the guards and makes sure that we start the keep alive
timer when it's really needed, and only stops it when we really know
that there is nothing left in the Windows message queue.

Task-number: QTBUG-17014
Reviewed-by: joao
Reviewed-by: Marius Storm-Olsen