Add a 'capabilities' flag to the marshaller and demarshaller
The marshaller and demarshaller need to know the features negotiated
with the peer so as to determine what's permitted. This is especially
important to the marshaller: the libdbus-1 API may "throw a fit" if we
try to pass a type that isn't allowed or the server may disconnect us.

The use of the capabilities in the demarshaller are for symmetry and
for us to toggle the feature in unit tests.

Task-number: QTBUG-17478