Fix crash when constantly referencing and dereferencing font engines
The font cache holds an implicit reference to font engines, garbage
collecting when an engine is unreferenced. We shouldn't delete font
engines from anywhere except the font cache, as this will lead to
a dangling pointer in the cache and it will inevitably crash when
you reference and dereference the same fonts a bunch of times.

Task-number: QTBUG-17423
Reviewed-by: Martin Zielinski