QScriptExtensionPlugin: fix crash when exiting autotest
The issue was caused by our InvalidValue() (created with
QScriptValue()) leaking some references. In my system this was caused
by two things acting together

- Plugins were not being unloaded properly, so their destructors
  were not called;
- Kross::EcmaPlugin when instantiated, created some QScriptValues in
  its Private object.

This commit changes availableExtensions() to unload the plugins
loaded, since we only want to get the keys. This avoids the issue for
our tests.

However the problem remains, if we add a simple QScriptValue member to
our test plugins, it will still break, because for importExtension, we
rely on the plugins being destroyed when application terminates, but
they are not, so InvalidValue() leaks again.

I've created a bug to investigate the issue with not destroying
plugins during termination:
http://bugreports.qt.nokia.com/browse/QTBUG-17458