Release scarce resources after evaluating bindings
Accessing any property in a binding that returns a scarce system
resource (such as a pixmap or texture) causes a JavaScript wrapper
object to be created which stores a copy of the scarce resource.
Since these wrapper objects are only deleted when the JavaScript
garbage collector is run, exposing any scarce system resource could
lead to exhausting system resources unless such scarce resources are
released automatically after the binding is evaluated.

This patch adds a new type of QDeclarativeScriptClass which allows
this sort of automatic resource releasing to occur, as well as the
ability for clients to explicitly preserve and release such resources
if required.

Task-number: QMLNG-18
Reviewed-by: Aaron Kennedy
Change-Id: I1c6d06b40ed67eb7a714b7c33800b78d3a7e3aaf