Add a mutex to protect the access to the QSet.
In theory, there should be no problems with doing this unlocked,
as the test should only run one thread at a time in this particular
code section.

In practice, if the test is failing, multiple threads would be
modifying the QSet. So the mutex is necessary to detect the
test failing.

Reviewed-By: Morten Sørvig