Fix thread starvation when processEvents() is called in a tight loop
One of the qftp test cases calls processEvents() in a tight loop instead
of using an event loop. This was causing the application to hang on symbian
as it blocked a lower priority system thread and the WLAN connection
never completed.
Although calling processEvents in a tight loop is bad practice, it works
on other OS where thread priorities are dynamic.

Reviewed-by: mread