Allow MouseArea to prevent mouse grab begin stolen by Flickable.
Placing a MouseArea in a Flickable (or PinchArea, or PathView) allows
the Flickable to steal mouse events when it detects a flick/drag event.
In some cases this is not desireable and MouseArea should be able to
retain its grab.  Added a 'preventStealing' property to prevent
an ancestor item from stealing the MouseArea's grab.

Change-Id: I6277fbb76919b2b35d4e32a247b38a90e305dbdf
Task-number: QTBUG-17285
Reviewed-by: Joona Petrell