Fix crash tst_QScriptExtQObject::connectAndDisconnect_senderWrapperCollected()
When in QtConnection::onSignal(), we previously used m_signal to
retreive the engine.

The fix is made of two parts. First, we make sure the QtConnexions are
deleted when their attached QtSignalData is deleted. Second, to ensure
that QObject-to-QObject connexions remain after their JS counterpart
is deleted, we connect the QObjects directly, i.e., we skip the
QtConnexion layer.

Reviewed-by: Olivier