Autotest: really ensure that two fds are equal
Instead of checking that they are both valid or both invalid, test
that they point to the same file on the filesystem. So we create a
QTemporaryFile and pass its file descriptor to the remote side and
back. If the two fds point to the same file on disk later (same st_dev
and st_ino), they are equal.

This probably works on non-Unix too, but I can't test and there's no
point anyway.

Task-number: QTBUG-17477