directfb: fix crash in client/server mode of qws
The server was crashing due to client pointer being sent as "sibling".

Instead surface flags are sent and sibling code is removed.

Also the window ID is transfered and used on receiver side to get a
local handle to the window.

Click to focus window works, but it is not raised.

Merge-request: 989
Reviewed-by: Marcel Schuette <marcel.schuette@nokia.com>