QScriptClass: implement support for name()
When intercepting getter calls for objects made by a QScriptClass, if
there isn't an existing "toString", we return our default
implementation: a function that returns name().

Since we are intercepting before the prototype() can "say" something,
we need to check if it has its own "toString", if so, we don't give
the default one.

We don't want to use Object default toString implementation, that's
why we get the "own property" of the prototype. This also excludes
other possible cases where user sets the toString of the prototype of
a prototype, but this seems to be a rare case. We can support later if
needed.

Reviewed-by: Olivier Goffart