Ensure animations start correctly when running is bound to a value.
Animations register to be started when the component is finalized,
but in the case of a binding, the registration doesn't happen until
we are already working with a copy of the finalization list. With this
patch, we no longer work with a copy, but with the actual list.

Change-Id: If460c250d2403590907e9ac854d277db68ba8b2a
Task-number: QTBUG-14042
Reviewed-by: Aaron Kennedy