Cocoa: implement eventdispatcher flag excludeUserInputEvents
When executing modal windows, the event dispatcher would not respect
the QEventLoop::ExcludeUserInputEvents. This patch implements
this support. The way it now works is that we fetch events manually
from NSApplication and do the dispatching ourselves in those cases.
From earlier research, this patch has shown to be a bit unreliable
in some cases, and might have some sideeffects. Hence the reason
we avoid it if we can. But it still feels better that we try to
follow the flag in those few application that wants to do this.