Cocoa: fix crash when using a drawer with a focus widget
The crash happends when you close down an application that has a
drawer that contains a widget with keyboard focus. The NSView
backing this widget is then the first responder in Cocoa.

This bug was a bit more hard-boiled than usual, and I can't say I truly
understand whats going on. My findings are that both the drawer and the
parent window both points to the same view inside the drawer as its
first responder. And when deleting the drawer (togheter with the focus
widget), the parent window is left pointing to a first responder that
is actually deleted. Is seems that us refusing to release a view as first
responder if we have no attached widget is wrong. So we choose to return
YES instead, which after all makes much more sense.

Task-number: QTBUG-15897
Reviewed-by: msorvig