Attempt to fix symbian 3.1 and 3.2 compile break.
e46c44f9538dbe5b44ce61d3a42403cfa471ae8b restructured qml.pri.  In the
original version the else part at the bottom of the file was processed
for Symbian 3.1 and 3.2, but the commit stopped that from happening.

This commit makes the minimal change to make the statements in the else
part apply to 3.1 and 3.2 again.

Really the file should be restructured to have separate logic for
setting each vairable, but I'll leave that task for the developers.

Reviewed-by: Eckhart Koppen