Cocoa: Sheets looses focus when moving parent window
The reason seems to be that we need to be more careful
what we answer when cocoa ask us if a sheet should get
main window status, and if the parent window should get
key status. So we add in two extra checks so we can
return the correct answer.

Task-number: QTBUG-15474
Reviewed-by: msorvig