Introduce QFontEngineDirectWrite
Make a font engine for subpixel positioned text on Windows Vista
(with platform update) and Windows 7. If selected during
configuration, the engine will be selected only when the hinting
preference of a font is set to None or Vertical hinting. The font
database uses most of the same logic but creates a direct write
font based on the LOGFONT rather than a GDI handle.

The engine is currently regarded as experimental, meaning that code
using it should do substantial testing to make sure it covers their
use cases.

Task-number: QTBUG-12678
Reviewed-by: Jiang Jiang