Fix duplicate calls to _q_startOperation
Because of the shared QNetworkSession, we need to disconnect the signals
before detaching from the session. Otherwise we may receive signals from
an old session after switching configurations.
Also, when a session is connected, we get both the state change (connected)
and the opened signals from the session. This needs to be distinguished
from the roaming->connected state change to avoid getting the
networkSessionConnected signal twice.

Reviewed-by: Markus Goetz
Task-Number: QTBUG-16901