QScriptContextPrivate: add way to customize callee
Add a way to provide a custom value to be returned when
QScriptContext::callee() is called in a native function context --
this value will also be used to fill the arguments object.

Before (which remains the default behavior) we got this information
from the v8::Arguments object. However, in callback invocations that
come because of SetCallAsFunctionHandler in an object, args.Callee()
is not the object being called as function.

It seems to me that args.Callee() has the "delegate function" used
internally by V8.

This addition to QScriptContextPrivate will allow us to create native
function contexts in which the callee() exposed in C++ API is the one
we expect.

Reviewed-by: Olivier Goffart