QScriptClass: implement Callable extension
Implement Callable extension by using function handler in
QScriptClassObject. Note that we don't use the wrapper helper for
invocation callback because we want to call a specific variant of
QScriptContextPrivate constructor (previous commit).

The extension_Callable tests all pass with one exception, which is the
thisObject() being "wrong". This is due how V8 implements function
handlers, details

    http://code.google.com/p/v8/issues/detail?id=1038

For callee() we could simply inject the value we wanted in QSC, but
for the thisObject() more magic is needed.

Reviewed-by: Olivier Goffart