QScriptValue: make isCallable() aware about QScriptClasses
QScriptClasses always register a function handler, but if they should
actually act as a callable depends on what the script class returns in
supportsExtension(QScriptClass::Callable).

This teaches isCallable() (which is used by QSVP::call()) to verify
the extension support in objects that have a QScriptClass.

With this tst_QScriptClass::extension_None now works as expected.

Reviewed-by: Olivier Goffart