DelayRemove of list delegate on section boundary duplicated section
When removing a delegate with a removal animation that fell on a
section boundary (i.e. owned the section header), the following
item would also create a section header before the previous item was
removed.  Make updateSections() include the removed, but visible items
in its update.  Ensure updateSections() is called when the removed
item is destroyed to ensure a new section header is created at that
point.

Change-Id: Ie831e3acf65b2989ebb030e2ab38cdbe179a9d45
Task-number: QTBUG-17606
Reviewed-by: Michael Brasser