Make autotest insensitive to exact JS error wording
This type of change was already done for the other QtScript
autotests. We don't guarantee a particular wording of messages,
since they can change between versions of the same back-end or
differ between back-ends. The only thing we care about in this
context is that a syntax error was in fact thrown.