Fix QPlainTextEdit corruption/crash with scrolling
There was an off-by-one error when QPlainTextEdit
calculated the number of visible lines (not blocks)
on the screne. When running into the error case,
the number of visible lines was falsely 0 (not 1),
which resulted in all softs of negative side effects
as the scrollbar permitted the user to scroll
behind the last paragraph.

Task-number: QTBUG-17230
Reviewed-by: con