Always prepend drive letter to $$EPOCROOT in Windows Symbian builds
Internal EPOCROOT handling in Qt already supports paths with and without
drive letter in both sbsv2 and abld builds, so we might as well make it
consistent and make sure the drive letter is prepended to $$EPOCROOT
value if it is missing.
This also makes paths deriving from $$EPOCROOT always usable in sbsv2
FLM files, which do not like paths without drive letters.

Task-number: QT-4611
Reviewed-by: Janne Koskinen