QRuntimeWindowSystem: Track window size properly
The runtime window system has a 'proxy' window surface which wraps the
currently active window surface. When the window geometry changes, the
new geometry is properly communicated to the wrapped window surface.
However, the new geometry is not updated into the runtime window surface
proxy itself, which means that when queried, the geometry for the window
surface will always be invalid. This patch fixes the issue.

Fixes: QT-4588

Merge-request: 1098
Reviewed-by: Samuel Rødal <samuel.rodal@nokia.com>