Add QT_NO_DYNAMIC_LIBRARY.
Set it for nacl, replace the Q_OS defines vxworks
already has in place.

Motivation: Support static plugins on platforms
that does not support dynamic plugins.

Static plugin support is implemented in QPluginLoader,
which has a QLibraryPrivate d pointer. The easiest
way to untangle this seems to be to compile in
QLibrary and then disable the ports of it that uses
dlopen.

Reviewed-By: Harald Fernengel