QSortFilterProxyModel::reset() should invalidate.
This was broken by: d149a3faca9b97ce806249bc7ef73fe2f59589d5

The connection was removed in that commit because we did not want
the layoutChanged signal to be emitted.  So instead of calling
invalidate(), we call clearMapping() directly.
In order to do that, clear_mapping has been turned into a private slot

Task-number: QTBUG-17812
Reviewed-by: Gabriel