Add missing API shims to QScriptValue constructors
This is needed to ensure that the calls into JSC are safe,
e.g. with regards to reentrancy.
I was not able to construct a testcase, but several of our
autotests crash without this change if
COLLECT_ON_EVERY_ALLOCATION is set to 1 in Collector.cpp
(this forces a garbage collection every time an object is
allocated).

Task-number: QTBUG-17815
Reviewed-by: Olivier Goffart