Make QtScript support COLLECT_ON_EVERY_ALLOCATION define
JSC has a define in runtime/Collector.cpp that can be enabled
to force garbage collection on every allocation. This can be
useful when investigating GC-related issues.
When the define is enabled, GC callbacks will happen before the
QScriptEngine(Private) is completely initialized, so we need to
initialize some things earlier (nice cleanup, actually), and
add some guards in the marking callback.
All tests pass with define off and on.

Task-number: QTBUG-17781
Reviewed-by: Olivier Goffart