Drag over bounds errors when ListView has variable height content
There are two related problems:
- when dragging variable height content up beyond bounds items are
destroyed, causing the maximum extent to change, which in turn causes
skipping.  This is fixed by storing the extents on press so that later
changes do not affect drag behavior.
- when variable height content is dragged beyond the bounds and
and released, items are created as the view returns to bounds.  This
changes the extents, however the bound target was not updated.  If the
viewport height changes during fixup the bounds animation is now
updated.

Change-Id: Ib37ec9e51f8ceb71af1e2e58f25fd8da18dcd632
Task-number: QTBUG-17769
Reviewed-by: Michael Brasser