Avoid asserting when computing line number for backtrace
With JSC asserts enabled (QtScript built without NDEBUG
defined), JSC::CodeBlock::getBytecodeIndex() would assert
because we sometimes called it with an address that was not
inside the range of the block's JIT code.

We never caught this bug because it just so happens that
even though the assert fails, the function returns a
result that causes our autotests to pass.

Check that the returnPC is in range and report lineNumber
-1 if not; this unifies the behavior of the interpreter
and JIT, even though it's not the result we want.

Task-number: QTBUG-17741
Reviewed-by: Olivier Goffart