Don't assert in abortEvaluation() autotest
When QtScript is built without NDEBUG defined, the
tst_QScriptEngine::abortEvaluation() test would assert.

This was due to commit
716e0284c8f569d71e42354fd6fc3b965233e019, which fixed
the tst_QScriptEngine::throwErrorFromProcessEvents()
autotest for a script containing an infinite while-loop
with an empty body.

The CHECK_FOR_EXCEPTION_AT_END() that we added should only
be done if the timeout checker did not report a timeout;
otherwise the JSC state becomes corrupted due to
returnToThrowTrampoline() being called twice. This caused
an assert later when calculating the line number of the
exception.

Also add test cases for scripts with try-catch statements.
For abortEvaluation(), scripts should not be able to
observe (i.e. catch) the interrupted exception, but if
an error is thrown using QScriptContext::throwError(), the
script should be able to catch it.

Task-number: QTBUG-17854
Reviewed-by: Olivier Goffart