Lighthouse: Make Native Interface slightly more agile
We realised that we couldn't keep on adding virtual functions for all
specific needs so we made one virtual function which takes a string
parameter which enables the implementation to return anything.

It seems a bit odd to have a virtual function as an accessor for a class
with only 1 virtual function. But the idea is that we don't want a
void * in the core lighthouse api. Also, the implementations should
decleare more functions to do the heavy lifting of getting the handles

Reviewed-by: paul