QScriptClass: Callable constructor (JS side)
Add support for an instance with a QScriptClass be called as
constructor. We stepped again on the limitations of v8 for
SetCallAsFunctionHandler(), in this case, it doesn't create the
'thisObject' for us.

This fixes the cases when the constructor is called using the new
operator in JavaScript.

Note that the test previously use a feature that doesn't work, so I've
added a property setting to check just if the created thisObject was
being correctly returned.

Reviewed-by: Jedrzej Nowacki
Reviewed-by: Olivier Goffart