v8: add a way to call an object function handler as constructor
This commit implements the Object::NewInstance() method, which allow
us call the object as a constructor (equivalent to "new obj()"). This
is done by _calling_ the function returned from GetConstructorDelegate().

Note that like v8, we don't implement the thisObject creation for
non-function objects called as constructors. The Qt code does that in
the handler itself when called as constructor. This is not ideal but
reduce the depth of changes in v8 code.

Reviewed-by: Olivier Goffart