Refactor qscriptv8testsuite to make it more maintainable
Move the infrastructure for creating a dynamic test object
to AbstractTestSuite, so that it can be reused by the
qscriptjstestsuite test as well (change coming in next commit).

Introduce configuration files for defining skipped tests and
expected failures; this was previously embedded in the C++ code,
which made it hard to update.

Make it possible to override the default test locations through
environment variables. This makes it easy to run the autotest
against an external repository (e.g. WebKit or V8 trunk), and
even different revisions of those repositories.

Task-number: QTBUG-17903
Reviewed-by: Jedrzej Nowacki