Move some network tests from gui.pro to network.pro
These network tests have no dependency on the QtGui dll, so I have moved
them to the network autotest group.
In some cases this was not stated in the tests' .pro files.

The gui autotest group still has a network test (qtcpsocket), but this
has a dependency on QtGui for one test case (which uses QMessageBox)

Reviewed-by: Markus Goetz