Don't assert in setProperty on a module API object
Previously, the QDeclarativeTypeNameScriptClass::setProperty()
function asserted if the object member was not set.  However, this
precondition is not valid in the case where the client is writing
to a property of a module API.

This commit moves the assert into the correct branch in setProperty().
Relates to commit ae6c7a44394d4a5f943f1fc92f4de015ec7bcc96.

Reviewed-by: Aaron Kennedy
Task-number: QMLNG-33
Change-Id: I4802da4c245f7b177609aef7f459c7adfc3b65c5