Support partial input mode
By default Symbian devices use fullscreen editing mode. However, in the
latest AVKON releases a partial virtual keyboard is supported. Support
the non-fullscreen editing mode from QCoeFepInputContext.

The non-fullscreen editing mode is only supported for Symbian^3 and
only for graphicsview-based solutions (QGraphicsView, QGraphicsWebView,
QDeclarativeView, ...). When native side indicates that the keyboard
opens, the graphicsview is possibly translated so that the text cursor
position is ensured to be visible. When keyboard closes, the
translation is removed. If the graphicsview contains vertical scrollbar,
the whole view is resized to the area above the keyboard, since
translating it, would move the upper part of the scrollbar out of
screen area.

There is a new exported private API to control when partial vkb is
used.

Task-number: QTBUG-16572
Reviewed-by: axis
Reviewed-by: Mrudul Pendharkar
Reviewed-by: Laszlo Agocs