Multiple screen support for Symbian in QDeskopWidget.
When using TV-out, the TV display can be used as an independent screen.
By default the content shown there is a clone of the device screen,
however from now on parenting a widget to QDesktopWidget::screen(1)
and calling show() will turn off cloning and have the widget shown
instead. screenCount() and the screenCountChanged signal can be used
to detect the availability of the secondary display, just like on other
platforms.

Task-number: QT-830

Reviewed-by: Sami Merila
Reviewed-by: Jani Hautakangas