Fix synchronous http deadlock when aborted by internal timeout
QHttpThreadDelegate::abortRequest was deleting itself, but not exiting
the event loop. For the synchronous usage, both these are incorrect.
Without exiting the event loop, the main thread waits forever.
If it deletes itself, then the main thread will access the deleted memory
on return (the delegate->incomingErrorCode class member) which can crash
with frequency depending on heap implementation.

With this change, abort acts more like the synchronousFinishedWithErrorSlot.

Reviewed-by: Markus Goetz