QDir: Fix absolute paths when using file engines
QFileSystemEntry doesn't know about paths as implemented by custom file
engines, such as is the case with resource files. In such cases, we need
to allow the engine to make the relative to absolute conversion.

Expanded QDir's test case to ensure resource paths are considered in
more places.

Task-number: QTBUG-17921
Reviewed-by: Prasanth Ullattil