QML app: text input field is not visible when split view is opened
When using a maximized QML application and user taps to a input widget,
opened splitview does not show the focused input widget.

The root cause of the problem is that when we move to splitview, we
set the window state to fullscreen. Unfortunately, there we were
trying to fullscreen the input widget and not the window that where
the input widget is.

Task-number: QTBUG-17984
Reviewed-by: Guoqing Zhang