Added automatic graphicssystem switching on meego when app is minimized.
When all top-level widgets are minimized we switch to raster to reduce
GPU memory consumption. We switch back to graphicssystem meego when at
least one top-level widget is shown normally again. The switching only
applies when the runtime graphicssystem is being used. The switching
only applies when the runtime graphicssystem is being used.

Task-number: QTBUG-18013
Reviewed-by: Armin Berres