Background app visible after split view closed
When using an app with graphicsview + vertical scrollbar that has
previously opened the splitview, when closing the keyboard, the
application is not reset to the original geometry correctly.

This is caused by invalid operation in the resetSplitView() method
that does not resize the graphics view back to original size AND
might even leave it with incorrect window state.

Task-number: QTBUG-17937
Reviewed-by: Guoqing Zhang