Fix network session use with the localhost optimisation in QNAB
QNetworkAccessBackend has an optimisation to not start the bearer when
the destination is localhost. On symbian, if the bearer is specified but
not started, then socket creation will fail.
To fix this, delay pushing the network session until start() is called,
at which point we know if the localhost optmisation will be applied or
not.
When using localhost, don't specify any network session - symbian socket
engine will create the socket successfully in thie case.

Reviewed-by: Markus Goetz