Regression: QS60Style: Theme background is incorrect
Fix for QTBUG-16816 cause theme background to display incorrectly
after a layout switch is done. Theme background is shown "tiled"
instead of being whole screen size.

This is because the setBackgroundTexture() call was removed to "save"
QPixmap creation. Unfortunately, it really skips the background
texture creation, but does not re-create a new one for the new layout,
so portrait background is used again. This leads to tiling the image.

Let's just return the call to the setBackgroundTexture()
(this will eventually anyway be "optimized" with fix for QTBUG-14910).

Task-number: QTBUG-17930
Reviewed-by: Laszlo Agocs