QS60Style: use placeholder texture when polishing widgets and palette
Background texture is not created until it is actually painted.
This allows skipping the whole background texture creation, if app
overwrites the QPalette::Window with its own data (image or color).

When widget is drawn and style notices that the widget is still using a
placeholder (1*1 red QPixmap) texture, it creates the real texture and
uses that for drawing.

Note that accessing QPalette::Window will give the placeholder pixmap.
Which is then again replaced with the real texture, if it is drawn
through the qt_s60_fill_background().

Task-number: QTBUG-14910
Reviewed-by: Laszlo Agocs